+++ client/routes/Register/components/Component.jsx
@@ -83,7 +83,7 @@ class Register extends BaseComponent {
           <ValidatedInput type='email' name='email' placeholder={emailPlaceholder} validations='isEmail' required autoFocus />
           <ValidatedInput type='text' name='name' placeholder={namePlaceholder} required validations='minLength:2' maxLength='60' />
           <PasswordInput type='password' name='password' placeholder={passwordPlaceholder} required />
-          <LaddaButton ref='button' isDisabled isLoading={this.props.state.register.loading} onSubmit={this.onSubmit}>
+          <LaddaButton ref='button' name='register-btn' isDisabled isLoading={this.props.state.register.loading} onSubmit={this.onSubmit}>
             <FormattedMessage
               id='register.RegisterBtn'
               defaultMessage='Register'
@@ -92,7 +92,7 @@ class Register extends BaseComponent {
           <center>{errorMsg}</center>
         </SecureForm>
         <p>
-          <Link to='/login'>
+          <Link name='already-having-account-link' to='/login'>
             <FormattedMessage
               id='register.AlreadyHaveAnAccount'
               defaultMessage='Already have an account? Log in here'