+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.2] - 2016-09-23
+### Added
+- Brome tests
+- Cookie banner
+- CRUD: filters_wildcard
+- CRUD: uids
+- Components examples
+
 ## [0.0.1] - 2016-06-28
 ### Added
 - initial release